+++ packages/wonder-blocks-modal/src/theme/default.ts
@@ -1,0 +1,52 @@
+import {border, semanticColor, sizing} from "@khanacademy/wonder-blocks-tokens";
+
+const theme = {
+    /**
+     * Shared tokens
+     */
+    root: {
+        border: {
+            radius: border.radius.radius_040,
+        },
+    },
+    /**
+     * Building blocks
+     */
+    dialog: {
+        spacing: {
+            padding: sizing.size_160,
+        },
+    },
+    footer: {
+        color: {
+            border: semanticColor.core.border.neutral.subtle,
+        },
+    },
+    header: {
+        color: {
+            border: semanticColor.core.border.neutral.subtle,
+            secondary: semanticColor.text.secondary,
+        },
+        spacing: {
+            paddingBlockMd: sizing.size_240,
+            paddingInlineMd: sizing.size_320,
+            paddingInlineSm: sizing.size_160,
+            gap: sizing.size_080,
+            // The space between the title and dismiss button.
+            titleGapMd: sizing.size_160,
+            titleGapSm: sizing.size_320,
+        },
+    },
+    panel: {
+        spacing: {
+            gap: sizing.size_320,
+        },
+    },
+    closeButton: {
+        spacing: {
+            gap: sizing.size_080,
+        },
+    },
+};
+
+export default theme;